* commit '0754239dd5d77d09d0b58daf1e17ce3d060afea8':
  gitignore
  加油及洗车优化

+++ XiaoMa/Class/GasPaymentResultVC.h
@@ -15,9 +15,9 @@
 
 @property (nonatomic, strong) NSString *detailText;
 @property (nonatomic, assign) DrawingBoardViewStatus drawingStatus;
-@property (nonatomic, assign) int paidMoney;
-@property (nonatomic, assign) int couponMoney;
-@property (nonatomic, assign) int chargeMoney;
+@property (nonatomic, assign) CGFloat paidMoney;
+@property (nonatomic, assign) CGFloat couponMoney;
+@property (nonatomic, assign) NSInteger chargeMoney;
 @property (nonatomic, assign) GasCard *gasCard;
 @property (nonatomic, copy) void(^dismissBlock)(DrawingBoardViewStatus status);
 @property (nonatomic, weak) UIViewController *originVC;
